feat: qmae - add loss for latent space

+++ src/qmae_latent_extension/utils/wrap_empty_indices.py
@@ -17,7 +17,7 @@ class WrappedDataset(Dataset):
         data = {
             "images": x,
             "indices": torch.zeros(16 * 16 + 1, dtype=torch.int64),
-            "features": torch.zeros(self.img_embedding_dim, dtype=torch.int64),
+            "features": torch.zeros(self.img_embedding_dim, dtype=torch.float64),
             "is_past_domain": self.is_past_domain,
         }
